% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{arg_parser}
\alias{arg_parser}
\title{Create Parser Object}
\usage{
arg_parser(option_list = NULL, prefix = "-", include_help = TRUE,
  description = "")
}
\description{
This function is a wrapper around the R6 Parser class to enable parser
creation and adding arguments/options through piping.
}

