% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-argument.R
\name{set_argument}
\alias{set_argument}
\title{Set Argument}
\usage{
set_argument(store_name, help_string = "", default = NULL, type = NULL,
  choices = NULL, n_args = 1)
}
\description{
This function creates instantiates a member of the argument class.
}

