% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-option.R
\name{set_option}
\alias{set_option}
\title{Set Option}
\usage{
set_option(option_strings, store_name = NULL, prefix = "-",
  help_string = "", default = NULL, is_flag = FALSE, type = NULL,
  choices = NULL, n_args = 1)
}
\description{
This function creates instantiates a member of the option class.
}
\keyword{internal}

