% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{data}
\name{Parser}
\alias{Parser}
\title{Create Argument Parser}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Parser
}
\description{
This function produces the argument parsing object.
}
\examples{

args <- arg_parser() \%>\%
   add_option(c('--verbose', '-v'), is_flag = TRUE, help = 'Prints verbose output.') \%>\%
   add_option('--multiply', type = 'numeric', help = 'Multiply by given number.') \%>\%
   parse_args()

}
\keyword{internal}

